% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tables-graphs.R
\name{bar_by_cat}
\alias{bar_by_cat}
\title{Render bar percentage}
\usage{
bar_by_cat(data, var, success = "Yes", .censor = TRUE, .gender_split = TRUE)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{var}{Variable to use}

\item{success}{Value(s) in \code{var} denoting 'success' (i.e. count as percentage)}

\item{.censor}{\code{TRUE}/\code{FALSE} - apply censoring rules (must be \code{TRUE} in output reports)}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}
}
\value{
A ggplot2 graph - percentage of successes in a single category.
}
\description{
Render bar percentage
}
