% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{bar_from_summary}
\alias{bar_from_summary}
\title{Bar percentage from summary data}
\usage{
bar_from_summary(summary_data)
}
\arguments{
\item{summary_data}{A dataframe produced by \code{create_collapsed_summary}}
}
\value{
A ggplot2 graph
}
\description{
Bar percentage from summary data
}
\examples{
N = 100
tibble(
  gender = sample(c("Girl", "Boy"), N, TRUE),
  class = sample(c("S1", "S6"), N, TRUE),
  answer = sample(c("Excellent", "Good", "Fair", "Poor"), N, TRUE)
) |>
  create_collapsed_summary(answer, success = c("Excellent", "Good")) |>
  bar_from_summary()
}
