% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mean-single.R
\name{bar_mean_single}
\alias{bar_mean_single}
\title{Bar graph of mean of a single var}
\usage{
bar_mean_single(summary_data, ymax, ylab = "Mean")
}
\arguments{
\item{summary_data}{A dataframe produced by \code{summary_mean_single}}

\item{ymax}{Upper limit of graph (deaults to \code{limits[2]})}

\item{ylab}{Label for X axis (summary statistic, i.e. "mean")}
}
\value{
A ggplot2 graph
}
\description{
Bar graph of mean of a single var
}
