% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tables-graphs.R
\name{bar_mean_single_var}
\alias{bar_mean_single_var}
\title{Bar graph of means for a single variable}
\usage{
bar_mean_single_var(
  data,
  var,
  .censor = TRUE,
  .gender_split = TRUE,
  limits = c(`Poor quality` = 1, `High quality` = 6),
  ymax = limits[2],
  ylab = "Mean"
)
}
\arguments{
\item{data}{Valid input data}

\item{var}{Variable to calculate mean of}

\item{.censor}{\code{TRUE}/\code{FALSE} - apply censoring rules (must be \code{TRUE} in output reports)}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}

\item{limits}{Vector of upper/lower score limits.}

\item{ymax}{Upper limit of graph (deaults to \code{limits[2]})}

\item{ylab}{Label for X axis (summary statistic, i.e. "mean")}
}
\value{
A ggplot2 bar graph of mean of a scoring variables
}
\description{
Bar graph of means for a single variable
}
