% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tables-graphs.R
\name{bar_multiple_vars}
\alias{bar_multiple_vars}
\title{Bar graph of \% categories for multiple variables}
\usage{
bar_multiple_vars(
  data,
  varslist,
  success = ~.x \%in\% c("More than once a week", "About every day"),
  group = c("gender", "class", "none"),
  .censor = TRUE,
  .gender_split = TRUE,
  classes = "All"
)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{varslist}{(named) list of variables to use. Names will provide graph labels}

\item{success}{A purr-like formula to determine observation in 'inclusion' group}

\item{group}{(Probably superseded) group bars by gender/class/none}

\item{.censor}{\code{TRUE}/\code{FALSE} - apply censoring rules (must be \code{TRUE} in output reports)}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}

\item{classes}{Vector names of classes to filter/combine.}
}
\value{
A ggplot2 graph of \% of 'success' in each group
}
\description{
Bar graph of \% categories for multiple variables
}
