% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions-multiple.R
\name{bar_proportions_multiple}
\alias{bar_proportions_multiple}
\title{Bar proportions multiple}
\usage{
bar_proportions_multiple(summary_data)
}
\arguments{
\item{summary_data}{Data produced by \code{summary_proportions_multiple}}
}
\value{
A ggplot2 graph
}
\description{
Bar proportions multiple
}
