% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tables-graphs.R
\name{bar_share_elevated}
\alias{bar_share_elevated}
\title{Bar graph of \% of sts with elevatved or expected mm scores}
\usage{
bar_share_elevated(data, .split, .censor = TRUE, classes = "All")
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{.split}{Split columns by gender x class}

\item{.censor}{\code{TRUE}/\code{FALSE} - apply censoring rules (must be \code{TRUE} in output reports)}

\item{classes}{Vector names of classes}
}
\value{
A ggplot2 graph
}
\description{
Bar graph of \% of sts with elevatved or expected mm scores
}
