% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_share-elevated-multiple.R
\name{bar_share_elevated_multiple}
\alias{bar_share_elevated_multiple}
\title{Produce bar graph of multiple  \% of sts with elevatved or expected scores}
\usage{
bar_share_elevated_multiple(graph_data)
}
\arguments{
\item{graph_data}{The output of \code{share_elevated_multiple}.}
}
\value{
A ggplot2 graph
}
\description{
Produce bar graph of multiple  \% of sts with elevatved or expected scores
}
