% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_addAge.R
\name{calculate_age}
\alias{calculate_age}
\title{Calculate age from date of birth and date of record}
\usage{
calculate_age(record_date, dobyr, dobmnth, dobday)
}
\arguments{
\item{record_date}{Date data collected (in format "\%Y-\%m-\%d \%H:\%M:\%S" or "\%d/\%m/\%Y \%H:\%M")}

\item{dobyr}{Year of birth}

\item{dobmnth}{Month of birth}

\item{dobday}{Day of birth}
}
\value{
An integer vector of ages
}
\description{
Calculate age from date of birth and date of record
}
