% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cleaningOutput.R
\name{cleaningOutput_server}
\alias{cleaningOutput_server}
\title{cleaningOutput Server}
\usage{
cleaningOutput_server(id, data)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{data}{reactive dataframe, from dataCleaning_server}
}
\description{
cleaningOutput Server
}
\keyword{internal}
