% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cleaningOutput.R
\name{cleaningOutputUI}
\alias{cleaningOutputUI}
\title{cleaningOutput UI}
\usage{
cleaningOutputUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
cleaningOutput UI
}
\keyword{internal}
