% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cleanUpload.R
\name{cleanUpload_server}
\alias{cleanUpload_server}
\title{cleanUpload Server}
\usage{
cleanUpload_server(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
cleanUpload Server
}
\keyword{internal}
