% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_cleanUpload.R
\name{cleanUploadUI}
\alias{cleanUploadUI}
\title{cleanUpload UI}
\usage{
cleanUploadUI(id, multiple = TRUE)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
cleanUpload UI
}
\keyword{internal}
