% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{create_collapsed_summary}
\alias{create_collapsed_summary}
\title{Collapsed summary (percentages of successes) for proportion graphs}
\usage{
create_collapsed_summary(
  data,
  var,
  success,
  genders,
  classes,
  .gender_split = FALSE
)
}
\arguments{
\item{data}{Valid input data}

\item{var}{Variable to calculate by}

\item{success}{Character vector of categories as 'successes'}

\item{genders}{List of genders to split by}

\item{classes}{List of classes to split by}

\item{.gender_split}{Gender split - passed from params}
}
\value{
A dataframe of proportions/counts of successes
}
\description{
Collapsed summary (percentages of successes) for proportion graphs
}
