% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_createReport_groupings.R
\name{createReport_groupings_server}
\alias{createReport_groupings_server}
\title{createReport_groupings Server}
\usage{
createReport_groupings_server(id, custom_group, report_type)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
createReport_groupings Server
}
\keyword{internal}
