% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_createReport_groupings.R
\name{createReport_groupingsUI}
\alias{createReport_groupingsUI}
\title{createReport_groupings UI}
\usage{
createReport_groupingsUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
createReport_groupings UI
}
\keyword{internal}
