% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_createReport.R
\name{createReport_server}
\alias{createReport_server}
\title{createReport Server}
\usage{
createReport_server(id, data)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{data}{reactive dataframe, from cleanUpload_server}
}
\description{
createReport Server
}
\keyword{internal}
