% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_createReport.R
\name{createReportUI}
\alias{createReportUI}
\title{createReport UI}
\usage{
createReportUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
createReport UI
}
\keyword{internal}
