% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_render-reports.R
\name{data_prep}
\alias{data_prep}
\alias{who_score}
\alias{mm_score}
\alias{sehs_primary}
\alias{sehs_secondary}
\title{Data preparation phase}
\usage{
data_prep(survey_data)

who_score(survey_data)

mm_score(survey_data)

sehs_primary(survey_data)

sehs_secondary(survey_data)
}
\arguments{
\item{survey_data}{The data to process}
}
\value{
\code{data_prep}: A dataframe with the required variables for rendering a report

\code{who_score}:  WHO 5-item wellbeing score (\code{who_score} variable) and categorical breakdown (\code{who_cat}: low/good)

\code{mm_score}: 'Me and My feelings' score
}
\description{
These functions prepare the data by reshaping it and generating necessary variables.
Helper functions for variables are included.
}
