% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataCleaning.R
\name{dataCleaning_server}
\alias{dataCleaning_server}
\title{dataCleaning Server}
\usage{
dataCleaning_server(id, data)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{data}{reactive dataframe, from rawUpload_server}
}
\description{
dataCleaning Server
}
\keyword{internal}
