% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataCleaning.R
\name{dataCleaningUI}
\alias{dataCleaningUI}
\title{dataCleaning UI}
\usage{
dataCleaningUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
dataCleaning UI
}
\keyword{internal}
