% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_population-tables.R
\name{date_range}
\alias{date_range}
\title{Generate text to describe input date range}
\usage{
date_range(input_data)
}
\arguments{
\item{input_data}{The dataframe of survey responses with the column \code{RecordedDate}}
}
\value{
A string describing month of survey (or range)
}
\description{
Generate text to describe input date range
}
