% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{geom_bar_t}
\alias{geom_bar_t}
\title{Thinner geom_bar}
\usage{
geom_bar_t(..., width = 0.7)
}
\arguments{
\item{...}{Other arguments to pass to \code{geom_bar}}

\item{width}{Width of bar (default 0.5)}
}
\description{
Thinner geom_bar
}
