% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_prep_hbsc.R
\name{get_hbsc_prop}
\alias{get_hbsc_prop}
\title{Get the HBSC Scotland proportion by gender for provided response}
\usage{
get_hbsc_prop(classes, success, var = NULL)
}
\arguments{
\item{classes}{classes to include}

\item{success}{responses to include, can provide multiple}

\item{var}{(Optional) variable to include}
}
\value{
A tibble of proportion by gender
}
\description{
Get the HBSC Scotland proportion by gender for provided response
}
