% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_prep_hbsc.R
\name{get_hbsc_score}
\alias{get_hbsc_score}
\title{Get the HBSC Scotland score by gender for provided response}
\usage{
get_hbsc_score(classes, var)
}
\arguments{
\item{classes}{classes to include}

\item{var}{variable to include}
}
\value{
A tibble of scores by gender
}
\description{
Get the HBSC Scotland score by gender for provided response
}
