% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_population-tables.R
\name{group_classes}
\alias{group_classes}
\title{Create a vector of nicely named, class groupings}
\usage{
group_classes(classes, groupings)
}
\arguments{
\item{classes}{character vector of classes}

\item{groupings}{list of groups of classes, e.g. list(c("S1", "S2", "S3"), c("S4", "S5", "S6"))}
}
\value{
A character vector of class groupings
}
\description{
Create a vector of nicely named, class groupings
}
