% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{how_many_valid}
\alias{how_many_valid}
\title{Return count of non-NA values}
\usage{
how_many_valid(x)
}
\arguments{
\item{x}{A variable with potentially NA values}
}
\value{
Count of non-NA values
}
\description{
Return count of non-NA values
}
