% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_checks.R
\name{make_upload_warning}
\alias{make_upload_warning}
\title{Make upload warning UI}
\usage{
make_upload_warning(message, level)
}
\arguments{
\item{message}{character}

\item{level}{integer, 1-3}
}
\value{
div html object
}
\description{
Make upload warning UI
}
\keyword{internal}
