% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_prep_hbsc.R
\name{prep_hbsc}
\alias{prep_hbsc}
\title{reshape hbsc summary data}
\usage{
prep_hbsc(dat = hbsc_scotland, create_cols = FALSE)
}
\arguments{
\item{dat}{HBSC data to pass in}

\item{create_cols}{copy existing p7, s2 and s4 columns to fill missing years}
}
\value{
long-format dataframe
}
\description{
reshape hbsc summary data
}
