% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pri_valid_responses}
\alias{pri_valid_responses}
\title{Simulated responses for primary survey testing}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 120 rows and 95 columns.
}
\usage{
pri_valid_responses
}
\description{
Simulated responses for primary survey testing
}
\keyword{datasets}
