% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{print_first_graph}
\alias{print_first_graph}
\alias{print_rest_of_graphs}
\title{Print multiple graphs across chunks}
\usage{
print_first_graph(graph_list)

print_rest_of_graphs(graph_list)
}
\arguments{
\item{graph_list}{List of graphs to print}
}
\value{
\code{print_first_graph} invisibly returns list of graphs with first dropped
}
\description{
\code{print_first_graph} prints the first graph and returns the remaining in a list
\code{print_rest_of_graphs} prints the rest
}
