% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{quiet_means}
\alias{quiet_means}
\title{Find the mean of only the numeric entries, no warnings}
\usage{
quiet_means(x)
}
\arguments{
\item{x}{A variable with mixed data types}
}
\value{
The mean of all numeric variables
}
\description{
Find the mean of only the numeric entries, no warnings
}
