% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_rawUpload.R
\name{rawUpload_server}
\alias{rawUpload_server}
\title{rawUpload Server}
\usage{
rawUpload_server(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
rawUpload Server
}
\keyword{internal}
