% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_render-reports.R
\name{render_report}
\alias{render_report}
\title{Render a single report for school or local authority}
\usage{
render_report(
  survey_data = NULL,
  school_name = NULL,
  local_authority_name = NULL,
  number_invited = NULL,
  output_location = getwd(),
  filename = "primary_report.docx"
)
}
\arguments{
\item{survey_data}{A clean dataframe of survey results}

\item{school_name}{If a school report, the school's name}

\item{local_authority_name}{If a local authority report, the LA's name (note not compatible with school report/name argument)}

\item{number_invited}{Number invited to complete the survey}

\item{output_location}{Location of file output (defaults to working directory)}

\item{filename}{Name of file to output}
}
\description{
This function renders a report from a given data file.
}
