% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{scale_colour_hbsc}
\alias{scale_colour_hbsc}
\title{Standard categorical fill scale}
\usage{
scale_colour_hbsc(...)
}
\arguments{
\item{...}{Other arguments passed to \code{scale_fill_manual}}
}
\description{
Standard categorical fill scale
}
