% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sec_valid_responses}
\alias{sec_valid_responses}
\title{Simulated responses for secondary survey testing}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 129 columns.
}
\usage{
sec_valid_responses
}
\description{
Simulated responses for secondary survey testing
}
\keyword{datasets}
