% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_share-elevated.R
\name{share_elevated}
\alias{share_elevated}
\title{Produce table of \% of sts with elevatved or expected mm scores}
\usage{
share_elevated(
  data,
  outcome,
  levels = c("As expected", "Elevated"),
  .split = TRUE,
  classes = "All",
  genders = c("Boy", "Girl")
)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{outcome}{The variable to graph}

\item{levels}{Levels of the variable (in ascending order)}

\item{.split}{Split columns by gender x class}

\item{classes}{Vector names of classes}

\item{genders}{Vector names of genders}
}
\value{
dataframe
}
\description{
Produce table of \% of sts with elevatved or expected mm scores
}
