% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_share-elevated-multiple.R
\name{share_elevated_multiple}
\alias{share_elevated_multiple}
\title{Proportions in each subgroup with elevated results across multiple vars}
\usage{
share_elevated_multiple(
  data,
  varlist,
  levels = c("As expected", "Elevated"),
  .split = TRUE,
  classes = "All",
  genders = c("Boys", "Girls")
)
}
\arguments{
\item{data}{Prepared input data}

\item{varlist}{List of variable labels, with names corresponding to columns}

\item{levels}{Levels to sum over}

\item{.split}{Split by gender/class}

\item{classes}{Vector/list of classes, nested by clusters}

\item{genders}{Vector of genders}
}
\value{
A table to plot
}
\description{
Proportions in each subgroup with elevated results across multiple vars
}
