% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mean-single.R
\name{summary_mean_single_var}
\alias{summary_mean_single_var}
\title{Summary table of means for a single variable}
\usage{
summary_mean_single_var(
  data,
  var,
  genders = c("Boys", "Girls"),
  classes = "All",
  .gender_split = TRUE
)
}
\arguments{
\item{data}{Valid input data}

\item{var}{Variable to calculate mean of}

\item{genders}{List of genders to split by}

\item{classes}{List of classes to split by}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}
}
\value{
A summary table of means of a single variable
}
\description{
Summary table of means for a single variable
}
