% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions-multiple.R
\name{summary_proportions_multiple}
\alias{summary_proportions_multiple}
\title{Table of \% categories for multiple variables}
\usage{
summary_proportions_multiple(
  data,
  varslist,
  success = ~.x \%in\% c("More than once a week", "About every day"),
  genders = c("Boys", "Girls"),
  classes = "All",
  .gender_split = TRUE
)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{varslist}{(named) list of variables to use. Names to match vars}

\item{success}{A purrr-like function denoting numerator categories}

\item{genders}{List of genders to split by}

\item{classes}{List of classes to split by}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}
}
\value{
A summary table of \% of 'success' in each group
}
\description{
Table of \% categories for multiple variables
}
