% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_population-tables.R
\name{tab_categories_grouped}
\alias{tab_categories_grouped}
\title{Table of grouped classes by gender counts}
\usage{
tab_categories_grouped(data, inc_gender, class_groupings)
}
\arguments{
\item{data}{Valid school input data (with columns \code{gender} and \code{class})}

\item{inc_gender}{List of genders to include in table}

\item{class_groupings}{List of groups of classes to include in table}
}
\value{
A flextable giving gender by grouped class counts
}
\description{
Table of grouped classes by gender counts
}
