% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_population-tables.R
\name{tab_responses}
\alias{tab_responses}
\title{Generate table reporting survey response rate}
\usage{
tab_responses(input_data, n_invited)
}
\arguments{
\item{input_data}{The dataframe of valid responses}

\item{n_invited}{Number invited to complete survey}
}
\value{
A table (markdown) detailing response rate
}
\description{
Generate table reporting survey response rate
}
