% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{table_from_summary}
\alias{table_from_summary}
\title{A table of percentages from summary data}
\usage{
table_from_summary(summary_data, inc_gender = genders)
}
\arguments{
\item{summary_data}{A dataframe produced by \code{create_collapsed_summary}}

\item{inc_gender}{List of genders to include in table}
}
\value{
A printed \code{flextable}
}
\description{
A table of percentages from summary data
}
\examples{

N = 100
tibble(
    gender = sample(c("Girl", "Boy"), N, TRUE),
    class = sample(c("S1", "S6"), N, TRUE),
    answer = sample(c("Yes", "No"), N, TRUE),
) |>
  create_full_summary(answer, levels = c("Yes", "No")) |>
  table_from_summary(c("Boys", "Girls", "All"))

}
