% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_checks.R
\name{upcheck_status}
\alias{upcheck_status}
\title{Data upload checks}
\usage{
upcheck_status(data)
}
\arguments{
\item{data}{}
}
\value{
single row data.frame of columns fail (logical), message (character), level (integer)
}
\description{
Data upload checks
}
\keyword{internal}
