% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_checks.R
\name{upload_checks}
\alias{upload_checks}
\title{Run all upload checks for raw data}
\usage{
upload_checks(data)
}
\arguments{
\item{data}{}
}
\value{
dataframe of failed checks
}
\description{
Run all upload checks for raw data
}
\keyword{internal}
