% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_checks.R
\name{upload_checks_raw}
\alias{upload_checks_raw}
\title{Run all upload checks for raw data or for clean data}
\usage{
upload_checks_raw(data, vars)
}
\arguments{
\item{data}{}
}
\value{
dataframe of failed checks
}
\description{
Run all upload checks for raw data or for clean data
}
\keyword{internal}
