% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graph-helpers.R
\name{valid_numbers}
\alias{valid_numbers}
\title{Return only numeric values, no warnings}
\usage{
valid_numbers(x)
}
\arguments{
\item{x}{A variable with mixed data types}
}
\value{
A vector of only valid numbers (as numeric)
}
\description{
Return only numeric values, no warnings
}
