% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_rules.R
\name{validators}
\alias{validators}
\alias{duration_too_short}
\alias{no_test_responses}
\alias{duplicate_cases}
\title{Validators}
\usage{
duration_too_short(data)

no_test_responses(data)

duplicate_cases(data)
}
\arguments{
\item{data}{Raw input dataset}
}
\value{
A tibble with the same length as the input data, showing the validation
results for the corresponding rows in the input data. Columns:
\itemize{
\item \code{message}: error message or an empty string
}
}
\description{
Validators
}
