% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{bar_from_summary}
\alias{bar_from_summary}
\title{Bar percentage from summary data}
\usage{
bar_from_summary(
  summary_data,
  inc_gender = genders,
  inc_classes,
  hbsc_data = NULL
)
}
\arguments{
\item{summary_data}{A dataframe produced by \code{create_collapsed_summary}}

\item{inc_gender}{List of genders to include in table}

\item{inc_classes}{List of classes to include in table}
}
\value{
A ggplot2 graph
}
\description{
Bar percentage from summary data
}
