% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{bar_from_summary}
\alias{bar_from_summary}
\title{Bar percentage from summary data}
\usage{
bar_from_summary(summary_data, hbsc_data = NULL)
}
\arguments{
\item{summary_data}{A dataframe produced by \code{create_collapsed_summary}}

\item{hbsc_data}{A dataframe with national data to add as points}
}
\value{
A ggplot2 graph
}
\description{
Bar percentage from summary data
}
