% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_tables-graphs.R,
%   R/report_mean-multiple.R
\name{bar_mean_multiple_vertical}
\alias{bar_mean_multiple_vertical}
\alias{bar_mean_multiple_vars}
\title{Bar graph of means of multiple variables}
\usage{
bar_mean_multiple_vertical(summary_data, ymax, ylab = "Mean")

bar_mean_multiple_vars(summary_data, xmax, xlab = "Mean")

bar_mean_multiple_vertical(summary_data, ymax, ylab = "Mean")
}
\arguments{
\item{xmax, ymax}{Upper limit of graph (deaults to \code{limits[2]})}

\item{xlab, ylab}{Label for X axis (summary statistic, i.e. "mean")}
}
\description{
Bar graph of means of multiple variables
}
