% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tables-graphs.R
\name{bar_mean_multiple_vars}
\alias{bar_mean_multiple_vars}
\alias{bar_mean_multiple_vertical}
\title{Bar graph of means for multiple variables (horizontal)}
\usage{
bar_mean_multiple_vars(
  data,
  varslist,
  group = c("gender", "class", "none"),
  .censor = TRUE,
  .gender_split = TRUE,
  limits = c(`Poor quality` = 1, `High quality` = 6),
  xmax = limits[2],
  xlab = "Mean",
  classes = "All"
)

bar_mean_multiple_vertical(
  data,
  varslist,
  group = c("gender", "class", "none"),
  .censor = TRUE,
  .gender_split = TRUE,
  limits = c(`Poor quality` = 1, `High quality` = 6),
  ymax = limits[2],
  ylab = "Mean",
  classes = "All"
)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{varslist}{(named) list of variables to use. Names to match vars}

\item{group}{(Probably superceded) group bars by gender/class/none}

\item{.censor}{\code{TRUE}/\code{FALSE} - apply censoring rules (must be \code{TRUE} in output reports)}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}

\item{limits}{Vector of upper/lower score limits.}

\item{xmax, ymax}{Upper limit of graph (deaults to \code{limits[2]})}

\item{xlab, ylab}{Label for X axis (summary statistic, i.e. "mean")}

\item{classes}{Vector names of classes to filter/combine.}
}
\value{
A ggplot2 bar graph of means across several scoring variables
}
\description{
\code{bar_mean_multiple_vars} returns a horizontal bar graph.
\code{bar_mean_multiple_vertical} returns a vertical graph.
}
