% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mean-multiple.R
\name{bar_mean_multiple_vars}
\alias{bar_mean_multiple_vars}
\alias{bar_mean_multiple_vertical}
\title{Bar graph of means of multiple variables}
\usage{
bar_mean_multiple_vars(summary_data, xmax, xlab = "Mean")

bar_mean_multiple_vertical(summary_data, ymax, ylab = "Mean")
}
\arguments{
\item{summary_data}{Data produced by \code{summary_mean_multiple_vars}}

\item{xmax, ymax}{Upper limit of graph}

\item{xlab, ylab}{Label for X axis (summary statistic, i.e. "Mean")}
}
\description{
\code{bar_mean_multiple_vars} returns a horizontal bar graph.
\code{bar_mean_multiple_vertical} returns a vertical graph.
}
