% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_share-elevated.R
\name{bar_share_elevated}
\alias{bar_share_elevated}
\title{Produce bar graph of \% of sts with elevatved or expected mm scores}
\usage{
bar_share_elevated(graph_data)
}
\arguments{
\item{graph_data}{The output of \code{share_elevated}.}
}
\value{
A ggplot2 graph
}
\description{
Produce bar graph of \% of sts with elevatved or expected mm scores
}
