% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_proportions.R
\name{create_full_summary}
\alias{create_full_summary}
\title{Full summary counts across all categories}
\usage{
create_full_summary(data, var, levels, genders, classes, .gender_split = FALSE)
}
\arguments{
\item{data}{Valid input data}

\item{var}{Variable to calculate by}

\item{levels}{Character vector of ordered levels}

\item{genders}{List of genders to split by}

\item{classes}{List of classes to split by}

\item{.gender_split}{Gender split - passed from params}
}
\value{
A dataframe of counted variables
}
\description{
Full summary counts across all categories
}
