% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_data_prep.R
\name{data_prep}
\alias{data_prep}
\alias{who_score}
\alias{mm_score}
\alias{sehs_primary}
\alias{sehs_secondary}
\alias{asw_score}
\alias{sdq_score}
\alias{fas_score}
\title{Data preparation phase}
\usage{
data_prep(survey_data, report_type = "primary")

who_score(survey_data)

mm_score(survey_data)

sehs_primary(survey_data)

sehs_secondary(survey_data)

asw_score(survey_data)

sdq_score(survey_data)

fas_score(survey_data)
}
\arguments{
\item{survey_data}{The data to process}

\item{report_type}{The type of survey data ("primary" or "secondary")}
}
\value{
\code{data_prep}: A dataframe with the required variables for rendering a report

The score-calculating functions return the dataset with relevant columns appended:

\code{who_score}:  WHO 5-item wellbeing score (\code{who_score} variable) and categorical breakdown (\code{who_cat}: low/good)

\code{mm_score}: 'Me and My feelings' score for primary schools

\code{sehs_primary}: SEHS score for primary schools

\code{sehs_secondary}: SEHS score for secondary schools

\code{asw_score}: Adolescent sleep-wake score for secondary schools

\code{sdq_score}: SDQ score for secondary schools

\code{fas_score}: family affluence score (0-13)
}
\description{
These functions prepare the data by reshaping it and generating necessary variables.
Helper functions for variables are included.
}
