% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbsc_scotland}
\alias{hbsc_scotland}
\title{HBSC Scotland 2020 summary statistics}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.
}
\usage{
hbsc_scotland
}
\description{
For comparisons with school-specific data
}
\keyword{datasets}
