% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_render-reports.R
\name{render_report}
\alias{render_report}
\title{Render a single report for school or local authority}
\usage{
render_report(
  survey_data,
  survey_type,
  school_name = NA,
  local_authority_name = NA,
  cluster_label = NULL,
  term = NULL,
  number_invited = NULL,
  output_location = NULL,
  gender_split = TRUE,
  classes = NULL,
  filename = "primary_report.docx"
)
}
\arguments{
\item{survey_data}{A clean dataframe of survey results}

\item{survey_type}{The report type - 'primary'/'secondary'}

\item{school_name}{If a school report, the school's name}

\item{local_authority_name}{If a local authority report, the LA's name (note not compatible with school report/name argument)}

\item{cluster_label}{Type of school grouping, e.g. local authority, primary cluster. Only used for 'local authority' type reports}

\item{term}{Name of term to print on report}

\item{number_invited}{Number invited to complete the survey}

\item{output_location}{Location of file output (defaults to working directory)}

\item{gender_split}{To split report outputs by gender}

\item{classes}{Vector or nested list of classes (clustered by combinations)}

\item{filename}{Name of file to output}
}
\description{
This function renders a report from a given data file.
}
