% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_mean-multiple.R
\name{summary_mean_multiple_vars}
\alias{summary_mean_multiple_vars}
\title{Table of means for multiple variables}
\usage{
summary_mean_multiple_vars(
  data,
  varslist,
  genders = c("Boys", "Girls"),
  classes = "All",
  .gender_split = TRUE
)
}
\arguments{
\item{data}{The dataframe of valid responses}

\item{varslist}{(named) list of variables to use. Names to match vars}

\item{genders}{List of genders to split by}

\item{classes}{List of classes to split by}

\item{.gender_split}{\code{TRUE}/\code{FALSE} - split by gender when sufficient numbers of responses}
}
\value{
A summary table of means of multiple variables
}
\description{
Table of means for multiple variables
}
