% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_population-tables.R
\name{tab_categories}
\alias{tab_categories}
\title{Table of class by gender counts}
\usage{
tab_categories(data, inc_gender, inc_classes)
}
\arguments{
\item{data}{Valid school input data (with columns \code{gender} and \code{class})}

\item{inc_gender}{List of genders to include in table}

\item{inc_classes}{List of classes to include in table}
}
\value{
A flextable giving gender by class counts
}
\description{
Table of class by gender counts
}
