% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_graphs.R
\name{table_from_summary}
\alias{table_from_summary}
\title{A table of percentages from summary data}
\usage{
table_from_summary(summary_data)
}
\arguments{
\item{summary_data}{A dataframe produced by \code{create_collapsed_summary}}
}
\value{
A printed \code{flextable}
}
\description{
A table of percentages from summary data
}
\examples{

tibble(
    gender = sample(c("Girl", "Boy"), N, TRUE),
    class = sample(c("S1", "S6"), N, TRUE),
    answer = sample(c("Yes", "No"), N, TRUE),
) |>
  create_full_summary(answer) |>
  table_from_summary()

}
