% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators}
\alias{validators}
\alias{check_duplicate_dob}
\alias{nothing_is_wrong}
\alias{everything_sucks}
\alias{everything_is_fine}
\title{Validators}
\usage{
check_duplicate_dob(data)

nothing_is_wrong(data)

everything_sucks(data)

everything_is_fine(data)
}
\arguments{
\item{data}{Raw input dataset}
}
\value{
A tibble with the same length as the input data, showing the validation
results for the corresponding rows in the input data. Columns:
\itemize{
\item \code{message}: error message or an empty string
}
}
\description{
Validators
}
