% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validators}
\alias{validators}
\alias{check_duplicate_dob}
\alias{nothing_is_wrong}
\alias{everything_sucks}
\alias{everything_is_fine}
\title{Validators}
\usage{
check_duplicate_dob(data)

nothing_is_wrong(data)

everything_sucks(data)

everything_is_fine(data)
}
\arguments{
\item{data}{Raw input dataset}
}
\value{
A character vector of validation messages, in the same order as the rows in the input data
}
\description{
Validators
}
