% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_rules.R
\name{validators}
\alias{validators}
\alias{duration_too_short}
\alias{no_test_responses}
\alias{partial_cases}
\alias{duplicate_cases}
\alias{suggest_missing_class}
\alias{age_year_mismatch}
\alias{straightlining}
\alias{recurring_postcodes}
\alias{no_consent}
\alias{has_school_id}
\alias{valid_dob}
\alias{completed_outside_school_hours}
\alias{completed_at_weekend}
\title{Validators}
\usage{
duration_too_short(data)

no_test_responses(data)

partial_cases(data)

duplicate_cases(data)

suggest_missing_class(data)

age_year_mismatch(data)

straightlining(data)

recurring_postcodes(data)

no_consent(data)

has_school_id(data)

valid_dob(data)

completed_outside_school_hours(data)

completed_at_weekend(data)
}
\arguments{
\item{data}{Raw input dataset}
}
\value{
A tibble with the same length as the input data, showing the validation
results for the corresponding rows in the input data. Columns:
\itemize{
\item \code{message}: error message or an empty string
}
}
\description{
Validators
}
